pyproject.toml
[project]
name = "soat.eleven.fastfood.pagamento"
version = "0.1.0"
description = "Microservice for payment management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "loguru>=0.7.3",
    "mercadopago>=2.3.0",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.4",
    "pytomlpp>=1.1.0",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "ruff>=0.14.6",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
lint = "ruff check src --fix && black src && isort src"
typecheck = "pyright src"
run_dev = "fastapi dev src/main.py"

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["soat.eleven.fastfood.pagamento"]
combine_as_imports = true
